Continuing adding support to TAS for various kinds of requests.
Add class PrintJsonEscaped, start using in place of existing printJsonEscapedTo
methods on string classes. Reduces the number of special features required of
a string class.

Modify class Array to be a compile time fixed size array of T, and rename the old
Array to ArrayView, a wrapper around a C-style array of T of some run-time specified size.

Add classes AnyPrintable and AnyPrintableArray, and function PrintableCat with
multiple specializations based on the number of parameters. These open up the
option of assembling an error message on the fly as it is being printed rather
than requiring string allocation.