Replace all uses of AnyString with AnyPrintable, and remove AnyString.
Put alpaca_response.* functions into struct WriteResponse so that the compiler helps to enforce consistency between the .h and .cc files.

Modify ServerConnectionBase, ServerConnection and TinyAlpacaServer to work with the existing Ethernet3 library, rather than requiring any changes there.

Continue the effort to make a host version of Ethernet3 for validating the server on host.