Fixes from compiling with Arduino IDE and debugging TestAlpacaDiscoveryServer.ino.
It appears that pgm_read_dword_far doesn't exist when compiled with Arduino IDE,
yet I find it in the arduino ide's directory. Not sure why this is yet.

Fixed bugs in AlpacaDiscoveryServer and in Literal (wrote order of args to memcmp_P,
etc.).

Disable TAS_CHECK on Arduino again: the avr-gcc (C++ 11) compiler doesn't support
__VA_OPT__, and I've not got a good way to work around that yet.