Pull PrintableProgmemString out of inline_literal.h. (Not sure if that is the right name yet.)
Rework make_enum_to_string.py so that it uses TASLIT for string literals,
and generate methods that return the string form of an enum.

Support adding the "Connection: close" header on responses conditionally.

Move the progmem literals of literals.inc into literals.cc (from literals.h).
This is a precursor to maybe using TASLIT to generate the literals.
I may also want to rework PrintableProgmemString and Literal so that
we separate the holding of a progmem literal from the various kinds of
uses of that string (printing, case free comparison, etc.).

Add an experiment where match_literals.cc doesn't use literal_token.h,
and instead does explicit comparisons, with the intent of NOT using
RAM for the tables of LiteralTokens.