More efforts to debug networking failures and crashes.
Status:
1) found that a this->flush() call as instead client_.flush(),
so that the flushing of the internal buffer was omitted. Sigh.

2) Now crashing (mostly) when I ctrl-C the host program
get_configured_devices.py right after sending a GET request.
The ctrl-c causes the connection to be closed before any response
is sent. After the crash, it appears that the networking hardware
no longer works properly, even though we are able to initialize
the listener sockets.