Add alpaca_client.py, an wrapper around alpaca_http_client.py that hides its HTTP and JSON nature, making it a bit more like using the ASCOM API.
Move some parts of alpaca_http_client.py to alpaca_model.py, a new file, so I
can have a set of classes unrelated to HTTP, and hence the requests package;
these can then be shared readily with other scripts.

Add switch_info.py, the start of a tool to print the list of available switches
in the switch devices found.